# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Libenga API",
        "version": "0.1"
    },
    "paths": {
...
        "/client": {
            "get": {
                "tags": [
                    "client"
                ],
                "summary": "Affiche Tous Client se trouvant dans la base de données",
                "operationId": "5781c1d61750090ca79e785c2cce6999",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "post": {
                "tags": [
                    "client"
                ],
                "summary": "Enregistrer un nouveau client",
                "operationId": "e8a58c07bfc028129fe11403bdc975a9",
                "responses": {
                    "200": {
                        "description": "Enregistrement avec success"
                    },
                    "400": {
                        "description": "erreur lors de l'enregistremnt"
                    }
                }
            }
        },
        "/client/{idClient}": {
            "get": {
                "tags": [
                    "client"
                ],
                "summary": " Affiche tous les clients ayant l'Id recherché",
                "operationId": "ab6a0ab3f6b9f47b90d2d0b35789b395",
                "parameters": [
                    {
                        "name": "idClient",
                        "in": "path",
                        "description": "idClient designe l'Id recherché",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "visualisation du client recherché"
                    },
                    "400": {
                        "description": "Id Invalide"
                    },
                    "404": {
                        "description": "le client n'existe pas"
                    }
                }
            }
        },
...
    }
}